server/requirements.txt
fastapi
uvicorn
sqlalchemy
apscheduler
python-dotenv
requests
